requirements.txt
flask>=2.3.0
librosa>=0.10.0
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.18.0
Pillow>=9.5.0
requests>=2.31.0
SQLAlchemy>=2.0.0
soundfile>=0.12.1
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0